// src/app/features/missions/types/mission.ts
import type {User} from "../../profil/types/user.ts";

export interface Metier {
  id_metier: string;
  nom: string;
}

export interface Mission {
  id_offre: string;
  id_utilisateur: string;
  id_metier: string;
  titre: string;
  description: string;
  localisation: string;
  prix: number;
  date_offre: string;
  statut: boolean;
  is_accepted: boolean;
  utilisateur: User;
  metier: Metier;
}